# MRG, ENH: Use numba to speed up summarize_clusters_stc