--- a/doc/changes/latest.inc
+++ b/doc/changes/latest.inc
@@ -147,6 +147,8 @@ Changelog
 
 - When passing a list of `~mne.Evoked` objects to `~mne.viz.plot_compare_evokeds`, each evoked's ``.comment`` attribute will be used to label the trace. If ``.comment`` is empty, a 1-based index is assigned as the label by  `Richard Höchenberger`_
 
+- Speed up :func:`mne.stats.summarize_clusters_stc` using Numba by `Yu-Han Luo`_
+
 Bug
 ~~~