@yh-luo 2020-08-05
> 
> 
> Nice that a tiny and still readable change can make such a big difference!
> 
> Can you update `latest.inc` to mention this speedup?

Sure, I'll do that tomorrow (it's nighttime now in my timezone 😉 )